Factor out "non-state RPC" logic.
Move out the logic for getauthmessage and setauthsignature (which just
calls through to libxidauth) into a new NonStateRpc helper class.  This
logic will be shared from there in the future with a light-mode binary
that exposes those RPCs without requiring a Xaya Core connection or
full game state.
